requirements.txt
appdirs==1.4.4
asgiref==3.3.0
black==20.8b1
click==7.1.2
dataclasses==0.7
Django==3.0.7
djangorestframework==3.12.1
mypy-extensions==0.4.3
pathspec==0.8.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.23
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
